Fix gemspec for CLI
When running the application via the gem installation, .git is not
present, but the Gemfile is used and tries to load the gemspec.  This
causes a warning when git is installed, and an error when git is not
installed.  Wrapping the files check prevents it from running except at
build time.