Fix execution of the generator in a Docker container
In order to make bundler respect the Gemfile when we run the CLI, we
need to have both development and regular dependencies pre-resolved.
This is the point of the Gemfile.lock, so we need to commit it.  In the
Dockerfile we only install the gem, which only installs non-dev
dependencies, but bundler still needs the full resolution tree which is
in the Gemfile.lock.

Fixes #55